% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{check_githubViolation}
\alias{check_githubViolation}
\title{check github event that might violate the within perio restrictions}
\usage{
check_githubViolation(start, end)
}
\arguments{
\item{start}{A date/time}

\item{end}{A date/time}

\item{username}{A character. Github login username}
}
\value{
A list of violation events
}
\description{
check github event that might violate the within perio restrictions
}
\examples{
none
}
