% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{check_status}
\alias{check_status}
\title{Check student progress between start and end}
\usage{
check_status()
}
\arguments{
\item{start}{A date/time}

\item{end}{A date/time}
}
\value{
A list of report
}
\description{
Check student progress between start and end
}
\examples{
none
}
