% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_scan_results.R
\name{correct_scan_results}
\alias{correct_scan_results}
\title{Correct Scan Results}
\usage{
correct_scan_results(file, solutions_file, id_nr="student_number", groups=c("group001", "group002", "group003"))
}
\arguments{
\item{file}{character, the path/filename of the csv file containing the scan results}

\item{solutions_file}{character, the path/filename of the csv file containing the sample solutions}

\item{points_file}{character, the path/filename of the csv file containing the points for correct responses per exam question}
}
\value{
a data.frame
}
\description{
Reads the scan results of running FormScan on the answer sheets, and the sample solutions, corrects the exams
}
\details{
The...
}
\examples{
myfile <- system.file("inst/exdata/example_scan_results.csv", package="examr")
solutions <- system.file("inst/exdata/example_sample_solutions.csv", package="examr")
my_results <- correct_scan_results(myfile, solutions)

}
\author{
Ulrich Matter <umatter@protonmail.com>
}
