% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{exam_authentication}
\alias{exam_authentication}
\title{authentication}
\usage{
exam_authentication(type = "auth")
}
\arguments{
\item{type}{a character}
}
\value{

}
\description{
authentication
}
\examples{
none.
}
