% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_answersheet.R
\name{generate_answersheet}
\alias{generate_answersheet}
\title{Generate Answer Sheet}
\usage{
generate_answersheet(as_template=NULL, title, date=Sys.Date())
}
\arguments{
\item{title}{character string, the title of the course/exam}

\item{date}{character string, the date of the exam (by default the system date)}
}
\value{
the path/filename of the generated answer sheet
}
\description{
Based on a template, generates an excel workbooks as answer sheet for FormScanner
}
\details{
The...
}
\examples{
generate_answersheet(title="Econ 101")
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
