% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_exam.R
\name{generate_exam}
\alias{generate_exam}
\title{Generate Exam Files}
\usage{
generate_exam(x)
}
\arguments{
\item{author}{character string, the author of the exam}

\item{title}{character string, the title of the course/exam}

\item{date}{character string, the date of the exam (by default the system date)}

\item{output}{character string, indicating in what format the exam and solution should be generated. either 'markdown', 'html', 'pdf', 'odt' or 'docx'. Default:pdf}

\item{file}{character, the path/filename of the excel-sheet containing the questions to be read}

\item{year}{character vector indicating from which years questions should be selected, if empty (NULL) questions from all years are selected.}

\item{mock}{logical, if TRUE, only mock exam questions are selected (FALSE by default)}

\item{topic}{character vector, indicating which question topics should be selected ('topic' column), defaults to NULL}

\item{types}{list, indicates how many questions should be selected from each type, default: one each}

\item{type_headers}{list, indicating the headers to be used for each question type section (see example)}
}
\value{
a list of pandoc objects (the function also generates output to disk, see output)
}
\description{
Generates pandoc documents for exam and solutions
}
\details{
The...
}
\examples{
myfile <- "inst/exdata/exam_questions.xlsx"
generate_exam(author="Ulrich Matter", title="Econ 101", output="pdf", file = myfile)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
