% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_points_frame.R
\name{generate_points_frame}
\alias{generate_points_frame}
\title{Generate Points Frame}
\usage{
generate_points_frame(points)
}
\arguments{
\item{file}{character, the path/filename of the csv file containing the scan results}

\item{points}{named list, list with key-value pairs question_group-points (e.g., list(group001=1.5, group002=3, group003=2))}
}
\value{
a data.frame
}
\description{
Generates a data frame with dimensions No. of responses x No. of scanned exams, containing the respective no. of points per correct response.
}
\details{
The...
}
\examples{
myfile <- system.file("inst/exdata/example_scan_results.csv", package="examr")
points <- list(group001=1.5, group002=3, group003=2)
my_results <- generate_points_frame(myfile, solutions)

}
\author{
Ulrich Matter <umatter@protonmail.com>
}
