% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_question_text.R
\name{generate_question_text}
\alias{generate_question_text}
\title{Generate Question Text}
\usage{
generate_question_text(x)
}
\arguments{
\item{x}{list, result of parsing file with questions via read_questions/select_questions}

\item{type_headers}{list, indicating the headers to be used for each question type section (see example)}
}
\value{
a list of character strings (with markdown syntax)
}
\description{
Generates Rmd strings for a set of questions
}
\details{
The...
}
\examples{
myfile <- "inst/exdata/exam_questions.xlsx"
my_questions <- read_questions(myfile)
selection <-select_questions(my_questions)
generate_question_text(selection)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
