% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitter.R
\name{getLastActivityTimeOfAllRooms}
\alias{getLastActivityTimeOfAllRooms}
\title{Get last activity time of each room belong to the authenticated user}
\usage{
getLastActivityTimeOfAllRooms()
}
\value{
A list of each room's last activity time
}
\description{
Get last activity time of each room belong to the authenticated user
}
\examples{
none
}
