% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_results.R
\name{grade_results_ch}
\alias{grade_results_ch}
\title{Grade Exams (CH Scale)}
\usage{
grade_results_ch(exams_corrected)
}
\arguments{
\item{exams_corrected}{character, a data frame with the total points per student (output of correct_scan_results)}

\item{c}{numeric, discount multiplier for grades under 4}

\item{quant}{numeric, quantile defining the threshold for the top grade (6)}

\item{ratio64}{numeric, ratio of minimum no. of points for a 6 and minimum points for a 4}

\item{round}{numeric, to what grade fractions should the grades be rounded (either .1, .5, or NULL, defaults to NULL). if NULL, grades won't be rounded}
}
\value{
a data.frame with the grade per student
}
\description{
Grades exam results based on total points and some fix parameters following the CH grade scale (1-6)
}
\details{
grade= a + b*points, if fail : grade =  4 - c (points(4) - points)
}
\examples{
myfile <- system.file("inst/exdata/example_scan_results.csv", package="examr")
solutions <- system.file("inst/exdata/example_sample_solutions.csv", package="examr")
my_results <- correct_scan_results(myfile, solutions)
grades <- grade_results(my_results)

}
\author{
Ulrich Matter <umatter@protonmail.com>
}
