% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_questions.R
\name{read_questions}
\alias{read_questions}
\title{Read Exam Questions}
\usage{
read_questions(file)
}
\arguments{
\item{file}{character, the path/filename of the excel-sheet containing the questions to be read}

\item{year}{character vector indicating from which years questions should be selected, if empty (NULL) questions from all years are selected.}

\item{mock}{logical, if TRUE, only mock exam questions are selected (FALSE by default)}

\item{topic}{character vector, indicating which question topics should be selected ('topic' column), defaults to NULL}
}
\value{
a list
}
\description{
Reads an excel sheet with exam question of different types (T/F, MC, etc.)
}
\details{
The...
}
\examples{
myfile <- "inst/exdata/exam_questions.xlsx"
my_questions <- read_questions(myfile)
head(my_questions)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
