% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scan_results.R
\name{read_scan_results}
\alias{read_scan_results}
\title{Read Scan Results}
\usage{
read_scan_results(file, id_nr="student_number", groups=c("group001", "group002", "group003"))
}
\arguments{
\item{file}{character, the path/filename of the csv file containing the scan results}

\item{id_nr}{character, group variable indicating the file-id (i.e., students number), defaults to "student_number"}

\item{groups}{named list indicating the possible responses defaults to c("group001", "group002", "group003")}
}
\value{
a data.frame/tibble
}
\description{
Reads and formats the scan results of running FormScan on the answer sheets
}
\details{
The...
}
\examples{
myfile <- system.file("inst/exdata/example_scan_results.csv", package="examr")
my_results <- read_scan_results(myfile)
head(my_results)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
