% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_questions.R
\name{select_questions}
\alias{select_questions}
\title{Select Questions for Exam}
\usage{
select_questions(x)
}
\arguments{
\item{x}{list, result of parsing file with questions via read_questions}

\item{types}{list, indicates how many questions should be selected from each type (see example below), default: one each}
}
\value{
a list of questions (same format as x)
}
\description{
Selects randomly a number of questions from a list of questions. MC options are randomly ordered.
}
\details{
The...
}
\examples{
myfile <- "data/exam_questions.xlsx"
my_questions <- read_questions(myfile)
select_questions(my_questions)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
