% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_solutions_text.R
\name{generate_solution_text}
\alias{generate_solution_text}
\title{Generate Solution Text}
\usage{
generate_solution_text(x, groups=c("group001", "group002", "group003"))
}
\arguments{
\item{x}{list, result of parsing file with questions via read_questions/select_questions}

\item{type_headers}{list, indicating the headers to be used for each question type section (see example)}

\item{groups}{named list indicating the possible responses defaults to c("group001", "group002", "group003")}
}
\value{
a list of character strings (with markdown syntax)
}
\description{
Generates Rmd strings for a set of questions with indicated sample solutions
}
\details{
The...
}
\examples{
myfile <- "inst/exdata/exam_questions.xlsx"
my_questions <- read_questions(myfile)
selection <-select_questions(my_questions)
generate_solution_text(selection)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
