% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_results.R
\name{grade_results_ch}
\alias{grade_results_ch}
\title{Grade Exams (CH Scale)}
\usage{
grade_results_ch(exams_corrected, c=0.18, quant=0.93, ratio64=2.68, round=.5, minp_6 = NULL )
}
\arguments{
\item{exams_corrected}{a data frame (output of correct_scan_results) or numeric vector with the total points per student}

\item{c}{numeric, discount multiplier for grades under 4}

\item{quant}{numeric, quantile defining the threshold for the top grade (6)}

\item{ratio64}{numeric, ratio of minimum no. of points for a 6 and minimum points for a 4}

\item{round}{numeric, to what grade fractions should the grades be rounded (either .1, .5, or NULL, defaults to NULL). if NULL, grades won't be rounded}

\item{minp_6}{numeric minimum number of points for best grade (6), defaults to NULL (minp_6 is determined based on quant)}
}
\value{
a data.frame with the grade per student
}
\description{
Grades exam results based on total points and some fix parameters following the CH grade scale (1-6)
}
\details{
grade= a + b*points, if fail : grade =  4 - c (points(4) - points)
}
\examples{
myfile <- system.file("inst/exdata/example_scan_results.csv", package="examr")
solutions <- system.file("inst/exdata/example_sample_solutions.csv", package="examr")
my_results <- correct_scan_results(myfile, solutions)
grades <- grade_results(my_results)

}
\author{
Ulrich Matter <umatter@protonmail.com>
}
